Separate server contexts from shared contexts.
